Fix default class names by moving to transformer
As pointed out by @eush77, it was not possible to preserve
default class names added by mdast-html. This fixes that by
moving them from the compiling phase (which cannot easily
be hooked into) to the transformation phase (which can be
hooked into by later attached plugins).

Note that plug-ins still need to check for the existance
of `node.data.htmlAttributes.class`, and join it with a
space if it exists.

Closes GH-3.